Add files via upload
IR transmitter always emits the IR (Infrared) rays which have to be received by IR receiver. When there is an obstacle in front of IR sensor, the transmitted IR rays are reflected back to the IR receiver. The output of Op-Amp is high when IR receiver receives the reflected IR rays. This output of operational amplifier is connected to the RESET pin of 555 timer.

The output of 555 timer is low when input voltage applied at pin 4 of 555 timer is low.

When input voltage at pin 4 is high, then 555 timer produces the frequency which can be adjusted by the different value combinations of resistors R1, R2 and capacitor C3.

Obstacle Present → IR Receiver Receives IR Rays → Op-amp Output High → 555 RST Pin High → Speaker Produces the Sound.
No Obstacle → IR Receiver does not Receive the Rays → Op-amp Output Low → RST Pin Low → No Sound
The output of NE555 is filtered by the 1uF capacitor and fed to the speaker.

Many IR sensors are arranged in house at different locations. The outputs of all these sensors are given to RESET pin of common 555 timer. When any sensor detects the intruder, the reset pin of 555 timer becomes high and speaker will produce the sound.